LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-en.html

<!DOCTYPE html><!-- webkit-test-runner [ language=en ] -->
<html>
<head>
<script src="../../../../../resources/js-test.js"></script>
</head>
<body>
<script>
shouldBeEqualToString("navigator.language", "en-us");
</script>
</body>
</html>
